+++ regta_period/__init__.py
@@ -1,7 +1,7 @@
-from .periods import AbstractPeriod, Period, PeriodAggregation
 from .enums import Weekdays
+from .periods import AbstractPeriod, Period, PeriodAggregation
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 __all__ = [
     "AbstractPeriod",
     "Period",