+++ tests/test_aggregation.py
@@ -6,6 +6,7 @@ from regta_period import Period, PeriodAggregation
 def _assert(p: PeriodAggregation, dt):
     assert p.get_next(dt) == datetime(year=1970, month=1, day=1, hour=11, minute=0, second=0)
     assert p.get_interval(dt) == timedelta(hours=11)
+    assert p.is_timezone_in_use is False
 
 
 def test_direct_creation(unix: datetime):